Moved IsEscaped method
